pyproject.toml
[project]
name = "ai-student-agent"
version = "0.1.0"
description = "AI学生费曼学习系统"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "python-dotenv",
  "langchain",
  "langchain-openai",
  "langchain-community",
  "langgraph",
  "langchainhub",
  "tiktoken",
  "chromadb",
  "pydantic",
  "unstructured[all-docs]",
  "bs4",
  "pandas",
  "requests",
  "scikit-learn",
  "lark",
  "jq",
  "numpy",
  "pillow",
  "urllib3",
  "docx2txt",
  "packaging",
  "multimethod",
  "pypdf",
  "openai",
  "weaviate-client",
  "pytesseract",
  "pdf2image",
  "pdfminer.six",
  "pdfplumber",
  "pikepdf",
  "pypdfium2",
  "lxml",
  "nltk",
  "tabulate",
  "tqdm",
  "xlsx2csv",
  "yarl",
  "aiohttp",
  "aiosignal",
  "async-timeout",
  "attrs",
  "frozenlist",
  "idna",
  "multidict",
  "tavily-python",
  "streamlit",
  "networkx",
  "pyvis",
  "plotly",
  "neo4j",
  "py2neo",
  "pytest",
  "httpx",
  "redis>=6.4.0",
  "celery>=5.5.3",
  "flower>=2.0.1",
  "supabase>=2.16.0",
  "zhipuai==2.1.5.20250825",
  "opentelemetry-exporter-otlp-proto-http>=1.36.0",
  "opentelemetry-instrumentation>=0.57b0",
  "opentelemetry-instrumentation-fastapi>=0.57b0",
  "opentelemetry-instrumentation-requests>=0.57b0",
  "graphviz>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/feynman"]

